More specific error handling in executor.
The problem I was seeing is the location of some errors was being
swallowed due to the aggressive error catching. Now, only errors in the
actual #call method will be rescued. All errors that happen outside of
it will raise like normal.

For example, when adding instrumentation code for this, if you have an
error in that code, this use to swallow it and make it impossible to
figure out what was causing error, but now it will raise that normally.